feat: types folder
