+++ apps/bot/src/types/api/TSMP.ts
@@ -1,0 +1,33 @@
+/* 
+Copyright 2022 Team Evie
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+export type McMMORes = {
+    repair: number;
+    fishing: number;
+    axes: number;
+    swords: number;
+    powerLevel: number;
+    alchemy: number;
+    Herbalism: number;
+    mining: number;
+    error: boolean;
+    acrobatics: number;
+    woodcutting: number;
+    excavation: number;
+    unarmed: number;
+    archery: number;
+    taming: number;
+};
